// src/types.ts
export interface AppsItem {
  id: number;
  nombre: string;
  descrip: string;
  categoria: string;
  pago: string;
  img: string;
  url: string;
  bgdestacado: string;
}

export interface PromptsItem {
  id: number;
  descrip: string;
  categoria: string;
  img: string;
}

export interface MetaDescrip {
  title: string;
  description: string;
}
 